// src/models/LiveStreaming/LiveStreamingInfo.ts
export interface LiveStreamingInfo {
    Id: string;
    LiveStreamingName: string;
    Description: string;
    StartTime: string | null;
    EndTime: string | null;
    LiveStreamingFlowName: string;
    Platform: string;
    Url1: string;
    Url2: string;
    SceneImage: string;
    MarkerImage: string;
    AdsImage: string;
    CompositeImage: string;
    Created: string | null;
    Updated: string | null;
}